// sourcecode/BTrak.Packages/snova-widget-module/projects/project-components/src/lib/widget-module/dependencies/models/projectlist.ts
export class ProjectList {
    projectId: string;
    projectName: string;
    id: string;
    responsiblePerson: string;
    projectResponsiblePersonName: string;
    testSuiteCount: number;
    testRunCount: number;
    openTestRunCount: number;
    milestoneCount: number;
    openMilestoneCount: number;
    totalCount: number;
    reportCount: number;
    isArchived: boolean;
    reportsCount: number;
    casesCount: number;
}