// sourcecode/BTrak.Packages/snova-widget-module/projects/project-components/src/lib/widget-module/dependencies/models/workspaceList.ts
export class WorkspaceList {
    workspaceId: string;
    workspaceName: string;
    isArchived: boolean;
    timeStamp: any;
    roleNames: string;
    roleIds: string;
    editRoleIds: string;
    deleteRoleIds: string;
    selectedRoleIds: any;
    isEditable: boolean;
    canView: boolean;
    canEdit: boolean;
    canDelete: boolean;
    isDefault: boolean;
    isHidden: boolean;
    description: string;
    isListView: boolean;
    isAppsInDraft: boolean;
    parentId: string;
}