// sourcecode/BTrak.Packages/snova-widget-module/projects/project-components/src/lib/widget-module/dependencies/models/generic-form-submitted.model.ts
export class GenericFormSubmitted {
    
    genericFormSubmittedId: string;
    formId: string;
    formJson: string;
    uniqueNumber: string;
    isArchived: boolean;
    genericFormName: string;
    timeStamp: any;
    formData: string;
    customApplicationId: string;
    publicFormId: string;
    isFormReadonly: boolean;
    isAbleToLogin: boolean;
}