// sourcecode/BTrak.Packages/snova-widget-module/projects/project-components/src/lib/widget-module/dependencies/models/department-model.ts
import { SearchCriteriaInputModelBase } from './searchCriteriaInputModelBase';
export class DepartmentModel extends SearchCriteriaInputModelBase {
    departmentId: string;
    departmentName: string;
    createdDateTime: Date;
    createdOn: string;
    inActiveDateTime: Date;
    timeStamp: any;
    totalCount: number;
}